refactor: step2, step3 패키지 이름 변경
