+++ src/main/kotlin/calculator/presentation/InputView.kt
@@ -1,0 +1,6 @@
+package calculator.presentation
+
+object InputView {
+
+    fun readInput(): String = requireNotNull(value = readln().takeIf { it.isNotBlank() })
+}
