+++ src/main/kotlin/racing/domain/game/RaceGameResult.kt
@@ -1,0 +1,3 @@
+package racing.domain.game
+
+data class RaceGameResult(val name: String, val position: Int)
