+++ src/main/kotlin/racing/domain/generator/NumberGenerator.kt
@@ -1,0 +1,6 @@
+package racing.domain.generator
+
+interface NumberGenerator {
+
+    fun generate(): Int
+}
