+++ src/main/kotlin/racing/view/CarPositionResult.kt
@@ -1,0 +1,3 @@
+package racing.view
+
+data class CarPositionResult(val name: String, val position: Int)
