--- src/main/kotlin/step2/presentation/InputView.kt
@@ -1,6 +1,0 @@
-package step2.presentation
-
-object InputView {
-
-    fun readInput(): String = requireNotNull(value = readln().takeIf { it.isNotBlank() })
-}
