--- src/main/kotlin/step3/domain/game/RaceGameResult.kt
@@ -1,3 +1,0 @@
-package step3.domain.game
-
-data class RaceGameResult(val name: String, val position: Int)
