--- src/main/kotlin/step3/domain/generator/NumberGenerator.kt
@@ -1,6 +1,0 @@
-package step3.domain.generator
-
-interface NumberGenerator {
-
-    fun generate(): Int
-}
