--- src/main/kotlin/step3/domain/generator/RandomNumberGenerator.kt
@@ -1,12 +1,0 @@
-package step3.domain.generator
-
-import kotlin.random.Random
-import kotlin.random.nextInt
-
-object RandomNumberGenerator : NumberGenerator {
-
-    private val RANDOM: Random = Random(seed = 0)
-    private val RANDOM_RANGE: IntRange = 0 until 10
-
-    override fun generate(): Int = RANDOM.nextInt(range = RANDOM_RANGE)
-}
