--- src/main/kotlin/step3/view/CarPositionResult.kt
@@ -1,3 +1,0 @@
-package step3.view
-
-data class CarPositionResult(val name: String, val position: Int)
