+++ src/test/kotlin/racing/domain/formula/NotMoveFormula.kt
@@ -1,0 +1,6 @@
+package racing.domain.formula
+
+object NotMoveFormula : MoveFormula {
+
+    override fun move(): Boolean = false
+}
