+++ src/test/kotlin/racing/domain/generator/MockNumberGenerator.kt
@@ -1,0 +1,6 @@
+package racing.domain.generator
+
+class MockNumberGenerator(private val value: Int) : NumberGenerator {
+
+    override fun generate(): Int = value
+}
