--- src/test/kotlin/step3/domain/formula/NotMoveFormula.kt
@@ -1,6 +1,0 @@
-package step3.domain.formula
-
-object NotMoveFormula : MoveFormula {
-
-    override fun move(): Boolean = false
-}
