--- src/test/kotlin/step3/domain/generator/MockNumberGenerator.kt
@@ -1,6 +1,0 @@
-package step3.domain.generator
-
-class MockNumberGenerator(private val value: Int) : NumberGenerator {
-
-    override fun generate(): Int = value
-}
